+++ app/Http/Controllers/CartController.php
@@ -151,27 +151,4 @@ class CartController extends Controller
             ]
         );
     }
-
-    public function order(OrderRequest $request)
-    {
-        $userId = Auth::id();
-        $order = $this->cartService->order($request, $userId);
-        $orderId = $order->id;
-        $listOrder = $this->cartService->orderDetail($orderId, $userId);
-        $cart = $this->cartService->getCartByUser($userId);
-        $this->cartService->deleteCart($cart->id);
-        $request->session()->forget('cart');
-        $user = $this->cartService->updateUserInfo($userId, $request);
-        $address = $this->cartService->createUserAddress($userId, $request);
-        $user->total_vnd = $order->vnd;
-        $user->total_usd = $order->usd;
-
-        return view(
-            'layouts.order',
-            [
-                'order' => $listOrder,
-                'user' => $user,
-            ]
-        );
-    }
 }
