+++ app/Service/CartService.php
@@ -5,10 +5,7 @@ use App\InterfaceService\CartInterface;
 use App\Cart;
 use App\Product;
 use App\CartDetail;
-use App\Order;
-use App\OrderDetail;
 use App\User;
-use App\Address;
 use Auth;
 
 class CartService implements CartInterface
@@ -128,40 +125,6 @@ class CartService implements CartInterface
         return $cartDetail;
     }
 
-    public function order($request, $userId)
-    {
-        $order = new Order();
-        $order->user_id = $userId;
-        $order->vnd = $request->vnd;
-        $order->usd = $request->usd;
-        $order->quantity = $request->quantity;
-        $order->save();
-
-        return $order;
-    }
-
-    public function orderDetail($orderId, $userId)
-    {
-        $cart = Cart::where('user_id', $userId)
-            ->first();
-        $cartDetails = $cart->detail;
-        $orderDetails = [];
-        foreach ($cartDetails as $detail) {
-            $order = new OrderDetail();
-            $order->order_id = $orderId;
-            $order->product_id = $detail->product_id;
-            $order->quantity = $detail->quantity;
-            $order->usd = $detail->usd;
-            $order->vnd = $detail->vnd;
-            $order->discount_usd = $detail->discount_usd;
-            $order->discount_vnd = $detail->discount_vnd;
-            $order->save();
-            array_push( $orderDetails, $detail);
-        }
-
-        return $orderDetails;
-    }
-
     public function deleteCart($cartId)
     {
         $cartDetails = CartDetail::where('cart_id', $cartId)
@@ -174,34 +137,6 @@ class CartService implements CartInterface
         }
 
         return $cart;
-    }
-
-    public function updateUserInfo($userId, $request)
-    {
-        $user = User::findOrFail($userId);
-        if(!$user) abort('404');
-        $user->phone = $request->phone;
-        $user->full_name = $request->name;
-        return $user;
-    }
-
-    public function createUserAddress($userId, $request)
-    {
-        $address = new Address();
-        $address->user_id = $userId;
-        $address->address = $request->address;
-        $address->lat = $request->lat;
-        $address->lng = $request->lng;
-        $address->active = 1;
-        $address->save();
-        $oldAddress = Address::where('user_id', $userId)
-            ->get();
-        foreach ($oldAddress as $address) {
-            $addr = Address::findOrFail($address->id);
-            $addr->active = 0;
-        }
-
-        return $address;
     }
 }
